refactor: colocate Stanford I2V task variants
