+++ mteb/tasks/retrieval/zxx/__init__.py
@@ -2,8 +2,10 @@ from .lp_music_caps import LPMusicCapsMTTA2TRetrieval, LPMusicCapsMTTT2ARetrieva
 from .music_caps import MusicCapsA2TRetrieval, MusicCapsT2ARetrieval
 from .song_describer import SongDescriberA2TRetrieval, SongDescriberT2ARetrieval
 from .sound_descs import SoundDescsA2TRetrieval, SoundDescsT2ARetrieval
-from .stanford_i2v_retrieval import StanfordI2VRetrieval
-from .stanford_i2v_visual_retrieval import StanfordI2VVisualRetrieval
+from .stanford_i2v_retrieval import (
+    StanfordI2VRetrieval,
+    StanfordI2VVisualRetrieval,
+)
 from .urban_sound8k_retrieval import UrbanSound8KA2TRetrieval, UrbanSound8KT2ARetrieval
 from .vim_sketch_retrieval import VimSketchA2ARetrieval
 from .vsc2022_retrieval import VSC2022Retrieval
